And we never crashed again forever and ever...

+++ rust/kleksis/src/app.rs
@@ -370,7 +370,7 @@ impl<'a> eframe::App for TemplateApp<'a> {
                         if let Some(cpos) = curr_pos {
                             if let Ok(block) = current.get_block(&id) {
                                 let mut next_temp = current.clone();
-                                let sb_idx: usize = if cpos.1 < block.size.1 - cpos.1 { 0 } else { 1 };
+                                let sb_idx: usize = if (cpos.1 as i32) < (block.size.1 as i32) - (cpos.1 as i32) { 0 } else { 1 };
                                 if next_temp.apply_cmd(&ProgCmd::LineCut(id, CutDirection::Y, cpos.1)).is_ok() {
                                     let cols = [
                                         calc_gmedian(
@@ -419,7 +419,7 @@ impl<'a> eframe::App for TemplateApp<'a> {
                                             target,
                                         ),
                                     ];
-                                    let sb_idx: usize = if cpos.0 < block.size.0 - cpos.0 { 0 } else { 1 };
+                                    let sb_idx: usize = if (cpos.0 as i32) < (block.size.0 as i32) - (cpos.0 as i32) { 0 } else { 1 };
                                     dispatch_cmd!(
                                         id,
                                         vec![
